+++ src/Webform.js
@@ -779,6 +779,9 @@ export default class Webform extends NestedComponent {
    * @param {string} message - The message to show in the alert.
    */
   setAlert(type, message) {
+    if (!type && this.submitted) {
+      return;
+    }
     if (this.options.noAlerts) {
       if (!message) {
         this.emit('error', false);